認証に関する処理をlib/authentication.tsxにまとめた
